Fix problem with leftover input
Consider the following script:

n = int(input())
print('n', n)
for i in range(n):
	print(input())
print('fim')

And the input: 3\n1\n2\n3

If we introduce an error on line 2, run the script, then fix the error
and run the script again, the second run would read the input starting from
the second number in the input.

To fix that, we now read the whole input (possibly left over from the
previous execution) before starting a new execution.